kern: mac: Make mac_map() return ssize_t
The map callback within the MAC ops returns ssize_t, therefore the
mac_map() wrapper should return the same

+++ src/sys/include/os/mac.h
@@ -117,7 +117,7 @@ int mac_check_creds(struct proc *procp, struct mac_border *mbp);
  * Returns zero on success, otherwise a less than zero value
  * on failure.
  */
-int mac_map(struct mac_border *mbp, off_t off, size_t len, void **res, int flags);
+ssize_t mac_map(struct mac_border *mbp, off_t off, size_t len, void **res, int flags);
 
 /*
  * Acquire a specific border using an ID
